DatabaseSQL: Don't use IF NOT EXISTS with CREATE TABLE, fix trigger
Basically, because we're creating triggers that run as long as the CREATE statement returns true, and the CREATE statement returns true even if the table already exists, we get a problem. So now the statement _will_ return false if it already exists, and the triggers won't run. Simply call @createTable() with errors supressed if this is the desired behaviour.

Relatedly, one of the triggers for current timestamp was not correctly escaped, so if the table name had a comma, it wouldn't be processed correctly. This has been fixed.